+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * 【core   】      CollUtil.subtract增加空判定（issue#3605@Github）
 * 【core   】      优化DateUtil.format(Date date, String format)接口效率（pr#1226@Gitee）
 * 【csv    】      CsvWriter.writeBeans增加重载，可选是否写出表头（issue#IA57W2@Gitee）
+* 【core   】      BetweenFormatter支持自定义设置单位（pr#1228@Gitee）
 
 ### 🐞Bug修复
 * 【core   】      修复AnnotationUtil可能的空指针错误